requirements.txt
beautifulsoup4>=4.12.2
selenium>=4.18.1
webdriver-manager>=4.0.1
python-dotenv>=1.0.1
pytest>=7.4.4